# docs/swagger.yaml
basePath: /api/v1
definitions:
  image.VoidZone:
    properties:
      height:
        type: integer
      offset_x:
        type: integer
      offset_y:
        type: integer
      width:
        type: integer
    type: object
  model.ImagesContainer:
    properties:
      candidate:
        type: string
      diff:
        type: string
      reference:
        type: string
    type: object
  model.Project:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  model.Reference:
    properties:
      reference_id:
        type: string
    type: object
  model.ResultContainer:
    properties:
      images:
        $ref: '#/definitions/model.ImagesContainer'
      percentage:
        type: number
    type: object
  model.Test:
    properties:
      candidate_id:
        type: string
      id:
        type: string
      result:
        $ref: '#/definitions/model.TestResult'
    type: object
  model.TestContainer:
    properties:
      approved:
        type: boolean
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      reference_id:
        type: string
      tests:
        items:
          $ref: '#/definitions/model.Test'
        type: array
      void_zones:
        items:
          $ref: '#/definitions/image.VoidZone'
        type: array
    type: object
  model.TestResult:
    properties:
      id:
        type: string
      percentage:
        type: number
    type: object
  model.errorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  model.successResponse:
    properties:
      result:
        type: string
    type: object
host: 127.0.0.1:3333
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Jameson API
  version: "1.0"
paths:
  /container/{container}:
    get:
      consumes:
      - application/json
      operationId: set container by id
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TestContainer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: container by id
  /container/{container}/add/voidzone:
    post:
      consumes:
      - application/json
      description: string container_id
      operationId: add voidzone for reference
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      - description: body
        in: body
        name: voidzones_array
        required: true
        schema:
          items:
            $ref: '#/definitions/image.VoidZone'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.successResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: add voidzone for reference
  /container/{container}/approve:
    patch:
      consumes:
      - application/json
      description: string container_id
      operationId: set container_id
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.successResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: approve reference for container
  /container/{container}/change/reference:
    patch:
      consumes:
      - application/json
      description: string container_id
      operationId: set new_reference_container
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      - description: body
        in: body
        name: mock
        required: true
        schema:
          $ref: '#/definitions/model.Reference'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.successResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: set new reference for container
  /container/{container}/delete:
    delete:
      consumes:
      - application/json
      operationId: delete_container
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.successResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: all projects
  /container/{container}/perform/test:
    post:
      consumes:
      - application/json
      - multipart/form-data
      operationId: set perform_test
      parameters:
      - description: container_id
        in: path
        name: container
        required: true
        type: string
      - description: this is a test file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TestResult'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: perform test
  /containers:
    get:
      consumes:
      - application/json
      operationId: get_containers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TestContainer'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: all containers
  /image/{image}:
    get:
      consumes:
      - application/json
      - multipart/form-data
      description: get image by id
      operationId: set image
      parameters:
      - description: image
        in: path
        name: image
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: image
  /project/{project}/test/create:
    post:
      consumes:
      - application/json
      - multipart/form-data
      operationId: set new_container
      parameters:
      - description: project_id
        in: path
        name: project
        required: true
        type: string
      - description: test_name
        in: query
        name: test_name
        required: true
        type: string
      - description: this is a test file
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.TestContainer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create new test container
  /project/create:
    post:
      consumes:
      - application/json
      operationId: creat_project
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: create project
  /projects:
    get:
      consumes:
      - application/json
      operationId: get_projects
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Project'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: get all containers
  /test/{test}:
    get:
      consumes:
      - application/json
      description: get test by id
      operationId: get result test data
      parameters:
      - description: test
        in: path
        name: test
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResultContainer'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.errorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/model.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            type: string
        default:
          description: ""
          schema:
            type: string
      summary: get result test data
swagger: "2.0"